[balancer_hierarchy] rename
